# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "REHAU Nea Smart 2 — Bridge API",
    "description": "REST + SSE access to a REHAU Nea Smart 2.0 base station.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/programs/daily": {
      "get": {
        "tags": [
          "programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 10
                      },
                      "name": {
                        "type": "string"
                      },
                      "bits": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1
                        },
                        "minItems": 96,
                        "maxItems": 96
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "bits"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/programs/daily/{n}": {
      "get": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "fresh",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10
            },
            "in": "path",
            "name": "n",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 10
                    },
                    "name": {
                      "type": "string"
                    },
                    "bits": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 1
                      },
                      "minItems": 96,
                      "maxItems": 96
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "bits"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "programs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "bits": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1
                        },
                        "minItems": 96,
                        "maxItems": 96
                      }
                    },
                    "required": [
                      "bits"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "intervals": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "start": {
                              "type": "string",
                              "pattern": "^([01]?\\d|2[0-4]):[0-5]\\d$"
                            },
                            "end": {
                              "type": "string",
                              "pattern": "^([01]?\\d|2[0-4]):[0-5]\\d$"
                            }
                          },
                          "required": [
                            "start",
                            "end"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "intervals"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10
            },
            "in": "path",
            "name": "n",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 10
                    },
                    "name": {
                      "type": "string"
                    },
                    "bits": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 1
                      },
                      "minItems": 96,
                      "maxItems": 96
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "bits"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/programs/weekly": {
      "get": {
        "tags": [
          "programs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      },
                      "name": {
                        "type": "string"
                      },
                      "days": {
                        "type": "array",
                        "minItems": 7,
                        "maxItems": 7,
                        "items": [
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          }
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "days"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/programs/weekly/{n}": {
      "get": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "fresh",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            },
            "in": "path",
            "name": "n",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 5
                    },
                    "name": {
                      "type": "string"
                    },
                    "days": {
                      "type": "array",
                      "minItems": 7,
                      "maxItems": 7,
                      "items": [
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "days"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "programs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "monday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "tuesday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "wednesday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "thursday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "friday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "saturday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "sunday": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  }
                },
                "required": [
                  "monday",
                  "tuesday",
                  "wednesday",
                  "thursday",
                  "friday",
                  "saturday",
                  "sunday"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            },
            "in": "path",
            "name": "n",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 5
                    },
                    "name": {
                      "type": "string"
                    },
                    "days": {
                      "type": "array",
                      "minItems": 7,
                      "maxItems": 7,
                      "items": [
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "days"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/programs/daily/{n}/intervals": {
      "get": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10
            },
            "in": "path",
            "name": "n",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "intervals": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "start": {
                            "type": "string"
                          },
                          "end": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "start",
                          "end"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "intervals"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}